feat: 商品删除时，商品文档和向量数据同步删除

+++ src/main/java/com/myproj/code/entity/GoodsDocument.java
@@ -34,5 +34,5 @@ public class GoodsDocument {
      * 关联的商品
      */
     @TableField("goods_id")
-    private Integer goodsId;
+    private Long goodsId;
 }
